fix: clear size ratchet and Windows privacy test failures
